requirements.txt
streamlit==1.24.0
google-generativeai==0.1.0
PyPDF2==3.0.1
pandas==1.5.3
matplotlib==3.7.1
pillow==9.5.0
SpeechRecognition==3.10.0
googletrans==3.1.0a0
pydub==0.25.1
python-dotenv==1.0.0
numpy==1.24.4
httpx==0.13.3
#httpx==0.24.1